Rename EmailAdmin to AdminEmail
