+++ frontend/src/metabase/query_builder/components/VisualizationError/components/AdminEmail/AdminEmail.tsx
@@ -2,7 +2,7 @@ import { useSetting } from "metabase/common/hooks";
 import CS from "metabase/css/core/index.css";
 import QueryBuilderS from "metabase/css/query_builder.module.css";
 
-export const EmailAdmin = () => {
+export const AdminEmail = () => {
   const adminEmail = useSetting("admin-email");
 
   if (!adminEmail) {
