--- frontend/src/metabase/query_builder/components/VisualizationError/components/EmailAdmin/index.ts
@@ -1,1 +1,0 @@
-export * from "./EmailAdmin";
